requirements.txt
Flask==2.0.2
requests==2.26.0
tefas-crawler==0.3.3
